[AND-462] Create configurable StreamVideoPlayerContent composable.
